move test to expr folder
